refactor(demo): presets move to src/lib/presets.ts
Pure data out of the 1100-line page component; adding a preset is now
a data-only diff.
